fix: update music and refactor spotify card
