+++ src/const/constants.ts
@@ -9,4 +9,5 @@ export const TWITTER_ID = "@shinryuzz__";
 export const AUTHOR = { name: "shinryuzz", url: WEBSITE_URL };
 
 export const SPOTIFY_EMBED_ID =
-  "https://open.spotify.com/embed/track/4hCLAWaQXvcEpUbGgyfEsX?utm_source=generator";
+  "https://open.spotify.com/embed/track/37J0tSBhYuE3F8j1pB1Vwr?si=6ebcc969c48742b4";
+// "https://open.spotify.com/embed/track/4hCLAWaQXvcEpUbGgyfEsX?utm_source=generator";
